Make simple changes based on comments
