+++ app/models/amazon_agent_coordinator_worker.rb
@@ -1,5 +1,6 @@
-class AmazonAgentManagerWorker < MiqWorker
+class AmazonAgentCoordinatorWorker < MiqWorker
   require_nested :Runner
+
   self.required_roles = ['smartproxy']
 
   # Don't allow multiple workers to run at once
